# lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poker Tournament Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Tournament {
  id: string;
  series: string;
  date: string;
  day: string;
  time: string;
  event: string;
  guarantee?: string | null;
  entry?: string | null;
  entryAmount?: number | null;
  chips?: string | null;
  levels?: string | null;
  endOfReg?: string | null;
  eventNumber?: string | null;
  format?: string | null;
  lateReg?: string | null;
  gameType: string;
  isMultiDay: boolean;
  isTurbo: boolean;
  isBounty: boolean;
  isSatellite: boolean;
}